deploy script for lcfa v2
