+++ packages/contracts/contracts/LenderCommitmentForwarder/LenderCommitmentForwarderV2.sol
@@ -1,15 +1,15 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.8.0;
 
-import "./LenderCommitmentForwarder_U1.sol";
+import "./LenderCommitmentForwarder_U2.sol";
 
-contract LenderCommitmentForwarderAlpha is LenderCommitmentForwarder_U1 {
+contract LenderCommitmentForwarderV2 is LenderCommitmentForwarder_U2 {
     constructor(
         address _tellerV2,
         address _marketRegistry,
         address _uniswapV3Factory
     )
-        LenderCommitmentForwarder_U1(
+        LenderCommitmentForwarder_U2(
             _tellerV2,
             _marketRegistry,
             _uniswapV3Factory
